Correctly use the recipients language in emails for the body
Since we use the template for the emails the body sometimes was displayed in
the language of the user, who triggered the email, rather then the recipient.

In order to be able to do this correctly, we need to pass in the actual
language and use that, instead of the language from the template engine.

Fix #145