cloudsuite: Implement the document loading via CloudSuite.
In order to work, we need a group that is shared between wwwrun (for owncloud)
and loolwsd; let's call it loolwww.  Then:

cd <www>/owncloud/apps/documents
mkdir documents-tmp
chown chown wwwrun:loolwww documents-tmp
chmod g+ws documents-tmp

ownCloud takes care of copying the files to this shared location, and back (in
the future).  It will also need to take care of the cleanup, at the moment the
files are leaked.