Convert 'Office' view to 'Recent documents' view
Instead of getting allll the documents of libreoffice mime types, we now
only get the most recently accessed ones.

This will increase the speed of loading that view for obvious reasons
(because there are less number of documents to load). It will also
decrease the number of database queries: before we used to hit a new
database query for each mimetype, but now we do all of it in one single
query.