Possibly fix Gallery & files_mediaviewer compatibility issue.
This simple, no-harm change checks that `OCA.Files.fileActions` exists
before using it.  This seems to fix #102, the compatibility issue between the
files_mediaviewer app and the old Gallery app.  The edit allows
javascript to continue without throwing an exception if
`OCA.Files.fileActions` does not exist.  For some reason this happens
when both Gallery and files_mediaviewer are enabled in my installation
of OwnCloud.