present login name to oauth2 client, keep user uid in auth tokens
In our case, the login name does not correspond to users UIDs. As such, when authenticating via OAuth2, there is a mismatch between the user ID coming from the query string, and the current logged user UID. We use the User Name throughout the flow, but we keep the user UID in the auth tokens.